refactor: listeners
